scalde/cognito_confidential/requirements.txt
requests>=2.28.2
python-jose[cryptography]>=3.3.0
authlib>=1.2.0
